requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.12; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
async-timeout==5.0.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
certifi==2025.4.26; python_version >= '3.6'
click==8.2.1; python_version >= '3.10'
deprecation==2.1.0
emoji==2.14.1; python_version >= '3.7'
exceptiongroup==1.3.0; python_version >= '3.7'
fastapi==0.115.12; python_version >= '3.8'
frozenlist==1.7.0; python_version >= '3.9'
gotrue==2.12.0; python_version >= '3.9' and python_version < '4.0'
h11==0.16.0; python_version >= '3.8'
h2==4.2.0; python_version >= '3.9'
hpack==4.1.0; python_version >= '3.9'
httpcore==1.0.9; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx[http2]==0.28.1; python_version >= '3.8'
hyperframe==6.1.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
iniconfig==2.1.0; python_version >= '3.8'
joblib==1.5.1; python_version >= '3.9'
multidict==6.4.4; python_version >= '3.9'
numpy==2.2.6; python_version >= '3.10'
packaging==25.0; python_version >= '3.8'
pluggy==1.6.0; python_version >= '3.9'
postgrest==1.0.2; python_version >= '3.9' and python_version < '4.0'
propcache==0.3.2; python_version >= '3.9'
pydantic==2.11.6; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pygments==2.19.1; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pytest==8.4.0; python_version >= '3.9'
pytest-mock==3.14.1; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.0; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
realtime==2.4.3; python_version >= '3.9' and python_version < '4.0'
scikit-learn==1.7.0; python_version >= '3.10'
scipy==1.15.3; python_version >= '3.10'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.46.2; python_version >= '3.9'
storage3==0.11.3; python_version >= '3.9' and python_version < '4.0'
strenum==0.4.15; python_version < '3.11'
supabase==2.15.3; python_version >= '3.9' and python_version < '4.0'
supafunc==0.9.4; python_version >= '3.9' and python_version < '4.0'
threadpoolctl==3.6.0; python_version >= '3.9'
tomli==2.2.1; python_version >= '3.8'
typing-extensions==4.14.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn[standard]==0.34.3; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.0.5; python_version >= '3.9'
websockets==14.2; python_version >= '3.9'
yarl==1.20.1; python_version >= '3.9'
